evm: cleanup, remove atlas/
