+++ .gitignore
@@ -13,6 +13,7 @@
 *.dll
 *.so
 *.dylib
+.dccache
 
 # Build
 *.test
