+++ types/int.go
@@ -1,0 +1,34 @@
+package types
+
+import (
+	fmt "fmt"
+	math "math"
+	"math/big"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+const maxBitLen = 256
+
+// SafeInt64 checks for overflows while casting a uint64 to int64 value.
+func SafeInt64(value uint64) (int64, error) {
+	if value > uint64(math.MaxInt64) {
+		return 0, sdkerrors.Wrapf(sdkerrors.ErrInvalidHeight, "uint64 value %v cannot exceed %v", value, int64(math.MaxInt64))
+	}
+
+	return int64(value), nil
+}
+
+// SafeNewIntFromBigInt constructs Int from big.Int, return error if more than 256bits
+func SafeNewIntFromBigInt(i *big.Int) (sdk.Int, error) {
+	if !IsValidInt256(i) {
+		return sdk.NewInt(0), fmt.Errorf("big int out of bound: %s", i)
+	}
+	return sdk.NewIntFromBigInt(i), nil
+}
+
+// IsValidInt256 check the bound of 256 bit number
+func IsValidInt256(i *big.Int) bool {
+	return i == nil || i.BitLen() <= maxBitLen
+}
