+++ x/evm/types/config.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ethereum/go-ethereum/params"
 )
 
-// EVMConfig encapulates common parameters needed to create an EVM to execute a message
+// EVMConfig encapsulates common parameters needed to create an EVM to execute a message
 // It's mainly to reduce the number of method parameters
 type EVMConfig struct {
 	Params      Params
