+++ x/evm/types/utils.go
@@ -2,7 +2,6 @@ package types
 
 import (
 	"fmt"
-	"math/big"
 
 	"github.com/gogo/protobuf/proto"
 
@@ -12,8 +11,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 )
-
-const maxBitLen = 256
 
 var EmptyCodeHash = crypto.Keccak256(nil)
 
@@ -91,17 +88,4 @@ func BinSearch(lo, hi uint64, executable func(uint64) (bool, *MsgEthereumTxRespo
 		}
 	}
 	return hi, nil
-}
-
-// SafeNewIntFromBigInt constructs Int from big.Int, return error if more than 256bits
-func SafeNewIntFromBigInt(i *big.Int) (sdk.Int, error) {
-	if !IsValidInt256(i) {
-		return sdk.NewInt(0), fmt.Errorf("big int out of bound: %s", i)
-	}
-	return sdk.NewIntFromBigInt(i), nil
-}
-
-// IsValidInt256 check the bound of 256 bit number
-func IsValidInt256(i *big.Int) bool {
-	return i == nil || i.BitLen() <= maxBitLen
 }
